Refactor pkg packages
