+++ pkg/cri/server/instrumented_service.go
@@ -24,7 +24,7 @@ import (
 	"golang.org/x/net/context"
 	runtime "k8s.io/cri-api/pkg/apis/runtime/v1alpha2"
 
-	ctrdutil "github.com/containerd/containerd/pkg/containerd/util"
+	ctrdutil "github.com/containerd/containerd/pkg/cri/util"
 )
 
 // instrumentedService wraps service with containerd namespace and logs.
