+++ pkg/cri/util/image.go
